+++ servers/rendering/renderer_rd/renderer_compositor_rd.h
@@ -79,7 +79,7 @@ protected:
 	Map<RID, RID> render_target_descriptors;
 
 	double time;
-	float delta;
+	double delta;
 
 	static uint64_t frame;
 
@@ -99,7 +99,7 @@ public:
 	void finalize();
 
 	_ALWAYS_INLINE_ uint64_t get_frame_number() const { return frame; }
-	_ALWAYS_INLINE_ float get_frame_delta_time() const { return delta; }
+	_ALWAYS_INLINE_ double get_frame_delta_time() const { return delta; }
 	_ALWAYS_INLINE_ double get_total_time() const { return time; }
 
 	static Error is_viable() {
